#914 - Modularize project import/export
- Moved automation export to automation module
- Relocated prerequisite exporters to api-dao module
