+++ webanno-automation/pom.xml
@@ -45,6 +45,10 @@
     </dependency>
     <dependency>
         <groupId>de.tudarmstadt.ukp.clarin.webanno</groupId>
+        <artifactId>webanno-api-dao</artifactId>
+    </dependency>
+    <dependency>
+        <groupId>de.tudarmstadt.ukp.clarin.webanno</groupId>
         <artifactId>webanno-diag</artifactId>
     </dependency>
     <dependency>
